Trmm.hpp moved to blast/math/algorithm
