+++ include/blast/math/algorithm/Trmm.hpp
@@ -5,12 +5,68 @@
 #pragma once
 
 #include <blast/math/Matrix.hpp>
-#include <blast/math/dense/TrmmBackend.hpp>
+#include <blast/math/RegisterMatrix.hpp>
+#include <blast/math/algorithm/Gemm.hpp>
 #include <blast/system/Tile.hpp>
+#include <blast/system/Inline.hpp>
 
 
 namespace blast
 {
+    namespace detail
+    {
+        template <size_t KM, size_t KN, typename T, typename P1, typename P2, typename P3>
+        requires MatrixPointer<P1, T> && (P1::storageOrder == columnMajor) && MatrixPointer<P2, T> && MatrixPointer<P3, T>
+        BLAST_ALWAYS_INLINE void trmmLeftUpper_backend(size_t M, size_t N, T alpha, P1 a, P2 b, P3 c)
+        {
+            size_t constexpr TILE_SIZE = TileSize_v<T>;
+            static_assert(KM % TILE_SIZE == 0);
+
+            RegisterMatrix<T, KM, KN, columnMajor> ker;
+
+            if (KM <= M)
+            {
+                size_t j = 0;
+
+                for (; j + KN <= N; j += KN)
+                {
+                    ker.reset();
+                    ker.trmmLeftUpper(alpha, a, b(0, j));
+                    gemm(ker, M - KM, alpha, a(0, KM), b(KM, j));
+                    ker.store(c(0, j));
+                }
+
+                if (j < N)
+                {
+                    auto const md = KM, nd = N - j;
+                    ker.reset();
+                    gemm(ker, M, alpha, a, b(0, j), md, nd);
+                    ker.store(c(0, j), md, nd);
+                }
+            }
+            else
+            {
+                // Use partial save to calculate the bottom of the resulting matrix.
+                size_t j = 0;
+
+                for (; j + KN <= N; j += KN)
+                {
+                    auto const md = M, nd = KN;
+                    ker.reset();
+                    gemm(ker, M, alpha, a, b(0, j), md, nd);
+                    ker.store(c(0, j), md, nd);
+                }
+
+                if (j < N)
+                {
+                    auto const md = M, nd = N - j;
+                    ker.reset();
+                    gemm(ker, M, alpha, a, b(0, j), md, nd);
+                    ker.store(c(0, j), md, nd);
+                }
+            }
+        }
+    }
     /// @brief C = alpha * A * B + C; A upper-triangular
     ///
     template <typename ST, typename MT1, typename MT2, typename MT3>
@@ -38,15 +94,15 @@ namespace blast
         // i + 4 * TILE_SIZE != M is to improve performance in case when the remaining number of rows is 4 * TILE_SIZE:
         // it is more efficient to apply 2 * TILE_SIZE kernel 2 times than 3 * TILE_SIZE + 1 * TILE_SIZE kernel.
         for (; i + 2 * TILE_SIZE < M && i + 4 * TILE_SIZE != M; i += 3 * TILE_SIZE)
-            trmmLeftUpper_backend<3 * TILE_SIZE, TILE_SIZE>(
+            detail::trmmLeftUpper_backend<3 * TILE_SIZE, TILE_SIZE>(
                 M - i, N, alpha, ptr<aligned>(A, i, i), ptr<aligned>(B, i, 0), ptr<aligned>(C, i, 0));
 
         for (; i + 1 * TILE_SIZE < M; i += 2 * TILE_SIZE)
-            trmmLeftUpper_backend<2 * TILE_SIZE, TILE_SIZE>(
+            detail::trmmLeftUpper_backend<2 * TILE_SIZE, TILE_SIZE>(
                 M - i, N, alpha, ptr<aligned>(A, i, i), ptr<aligned>(B, i, 0), ptr<aligned>(C, i, 0));
 
         for (; i + 0 * TILE_SIZE < M; i += 1 * TILE_SIZE)
-            trmmLeftUpper_backend<1 * TILE_SIZE, TILE_SIZE>(
+            detail::trmmLeftUpper_backend<1 * TILE_SIZE, TILE_SIZE>(
                 M - i, N, alpha, ptr<aligned>(A, i, i), ptr<aligned>(B, i, 0), ptr<aligned>(C, i, 0));
     }
